22 April 2021

+++ app/Http/Controllers/AuthController.php
@@ -117,7 +117,7 @@ class AuthController extends Controller
     }
 
     public function logout(Request $request){
-        $request->user()->tokens()->delete();
+        $request->user()->tokens()->where('id', $request->user()->currentAccessToken()->id)->delete();
         $response = [
             'success'           =>  true,
             'message'           => 'Logout successful.'
